pyproject.toml
[project]
name = "Enterprise CarbonScope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.140.0",
    "crewai-tools>=0.49.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.5",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "xlsxwriter>=3.2.5",
]
